\name{haplodiplo-package}
\alias{haplodiplo-package}
\alias{haplodiplo}
\docType{package}
\title{
\packageTitle{haplodiplo}
}
\description{
\packageDescription{haplodiplo}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{haplodiplo}
\packageIndices{haplodiplo}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{haplodiplo}

Maintainer: \packageMaintainer{haplodiplo}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
